Add autoscroll, grid icon size scaling, theme compatibility fixes
Create unique bookmark arrays
Add a spacer to pad the toggle button from the search box
Fix chromium bookmark indexing, recurse bookmark structures
Add category icon size setting, theme introspection
Fix context menu behavior
Add overlay context menus for grid view
Add context menu functionality
Add context menu functionality
Move CinnamenuButton, CinnamenuPanel, buttons classes into separate
files
Rename PanelMenuButton to CinnamenuPanel
Fix list view app labels not appearing on LM theme, fix category button
hover CSS, narrow container width
