+++ Cinnamenu@json/CHANGELOG.md
@@ -2,5 +2,11 @@ Changelog
 
 ### 1.0.0
 
+  * Reimplemented the context menu.
+  * Overhauled theme compatibility.
+  * Added a (default) option to scale grid icons based on how many apps are in the list to make the best use of the screen space.
   * Porting from GNOME Shell is complete, and most of the functionality is restored from the original version.
   * Migrated most of the styling to Cinnamon's CSS.
+  * Todo:
+    * Add keyboard support.
+    * Clean up and organize existing localization files, and add new translations.
