+++ Cinnamenu@json/files/Cinnamenu@json/settings-schema.json
@@ -15,7 +15,7 @@
   },
   "header1": {
     "type": "header",
-    "description": "Panel Appearance"
+    "description": "Appearance"
   },
   "menu-icon-custom": {
     "type": "checkbox",
@@ -37,15 +37,15 @@
     "description": "Set the custom menu button icon name",
     "tooltip": "Set the custom menu button icon name"
   },
-  "shortcuts-icon-size": {
+  "category-icon-size": {
     "type": "spinbutton",
-    "default": 18,
+    "default": 24,
     "min": 16,
     "max": 256,
     "step": 1,
     "units": "pixels",
-    "description": "Shortcuts Panel icon size",
-    "tooltip": "Shortcuts Panel icon size"
+    "description": "Category icon size",
+    "tooltip": "Category icon size"
   },
   "apps-list-icon-size": {
     "type": "spinbutton",
@@ -66,6 +66,12 @@
     "units": "pixels",
     "description": "Applications grid icon size",
     "tooltip": "Applications grid icon size"
+  },
+  "apps-grid-icon-scale": {
+    "type": "checkbox",
+    "default": true,
+    "description": "Scale grid icons by available space",
+    "tooltip": "This makes the grid icon size variable depending on how many apps are being shown. This overrides the grid icon size option."
   },
   "apps-grid-label-width": {
     "type": "spinbutton",
@@ -107,6 +113,12 @@
     "default": false,
     "description": "Use menu animations",
     "tooltip": "Allow the menu to animate on open and close"
+  },
+  "enable-autoscroll": {
+    "type": "checkbox",
+    "default": true,
+    "description": "Enable autoscrolling in application list",
+    "tooltip": "Choose whether or not to enable smooth autoscrolling in the application list."
   },
   "startup-view-mode": {
     "type": "combobox",
